Prefs/Reqtools: Compiler delint

git-svn-id: https://svn.aros.org/svn/aros/trunk@48244 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/prefs/reqtools/reqtoolsstringify.c
@@ -47,7 +47,7 @@ static IPTR ReqToolsStringify__MUIM_Numeric_Stringify(Class  * cl, Object * obj,
 
     D(bug("[ReqToolsEditor.class] %s()\n", __PRETTY_FUNCTION__));
 
-    sprintf((char *)data->buf, "%3d%%", msg->value);
+    sprintf((char *)data->buf, "%3d%%", (int)msg->value);
 
     return (IPTR)data->buf;
 }
